Sync latest Nucleo firmware and timing architecture.
Integrate ISR-based 1 kHz sampling with 50 ms rolling-window inference, stabilize UART/diagnostic logging, and align project/docs with the current STM32U575 workflow and fake-data validation path.

+++ SensaLimbNucleoU575ZIQ/Core/Inc/adc_fake_data.h
@@ -4,7 +4,7 @@
 #include <stddef.h>
 #include <stdint.h>
 
-extern const uint16_t g_fake_samples[];
+extern const float g_fake_samples[];
 extern const size_t g_fake_frame_count;
 extern const size_t g_fake_channels;
